\name{MVSKFamily-package}
\alias{MVSKFamily-package}
\alias{MVSKFamily}
\docType{package}
\title{
\packageTitle{MVSKFamily}
}
\description{
\packageDescription{MVSKFamily}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{MVSKFamily}

Maintainer: \packageMaintainer{MVSKFamily}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = MVSKFamily)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{MVSKFamily}}
%\section{Documentation Index}{\packageIndices{MVSKFamily}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
